+++ src/main/java/com/bettercontent/proceduralbouquets/registry/ModMenus.java
@@ -1,0 +1,25 @@
+package com.bettercontent.proceduralbouquets.registry;
+
+import com.bettercontent.proceduralbouquets.ProceduralBouquets;
+import com.bettercontent.proceduralbouquets.menu.BouquetGridMenu;
+import net.minecraft.world.inventory.MenuType;
+import net.minecraftforge.common.extensions.IForgeMenuType;
+import net.minecraftforge.eventbus.api.IEventBus;
+import net.minecraftforge.registries.DeferredRegister;
+import net.minecraftforge.registries.ForgeRegistries;
+import net.minecraftforge.registries.RegistryObject;
+
+public final class ModMenus {
+    private static final DeferredRegister<MenuType<?>> MENUS =
+        DeferredRegister.create(ForgeRegistries.MENU_TYPES, ProceduralBouquets.MOD_ID);
+
+    public static final RegistryObject<MenuType<BouquetGridMenu>> BOUQUET_GRID =
+        MENUS.register("bouquet_grid", () -> IForgeMenuType.create(BouquetGridMenu::new));
+
+    private ModMenus() {
+    }
+
+    public static void register(IEventBus bus) {
+        MENUS.register(bus);
+    }
+}